+++ src/FSharp.Analyzers.SDK/FSharp.Analyzers.SDK.Client.fs
@@ -7,15 +7,21 @@ open System.Runtime.Loader
 open McMaster.NETCore.Plugins
 open System.Collections.Concurrent
 
+type AnalysisResult = {
+  AnalyzerName : string
+  Output : Result<Message list, exn>
+}
+
 module Client =
 
   let internal attributeName = "AnalyzerAttribute"
 
   let internal isAnalyzer (mi: MemberInfo) =
     mi.GetCustomAttributes true
-    |> Seq.exists (fun n -> n.GetType().Name = attributeName)
+    |> Seq.tryFind (fun n -> n.GetType().Name = attributeName)
+    |> Option.map unbox<AnalyzerAttribute>
 
-  let internal analyzerFromMember (mi: MemberInfo) : Analyzer option =
+  let internal analyzerFromMember (mi: MemberInfo) : (string * Analyzer) option =
     let inline unboxAnalyzer v =
       if isNull v then
         failwith "Analyzer is null"
@@ -30,15 +36,8 @@ module Client =
           then Some(m.Invoke(null, null) |> unboxAnalyzer)
         elif m.ReturnType.FullName.StartsWith "Microsoft.FSharp.Collections.FSharpList`1[[FSharp.Analyzers.SDK.Message" then
           try
-            let x : Analyzer = fun ctx ->
-              try
-                m.Invoke(null, [|ctx|]) |> unbox
-              with
-              | ex ->
-                printfn "Error while executing Analyzer from %s.%s" m.DeclaringType.Name m.Name
-                printfn "%A" ex
-                []
-            Some x
+            let analyzer : Analyzer = fun ctx -> m.Invoke(null, [|ctx|]) |> unbox
+            Some analyzer
           with
           | ex -> None
         else None
@@ -46,7 +45,13 @@ module Client =
         if m.PropertyType = typeof<Analyzer> then Some(m.GetValue(null, null) |> unboxAnalyzer)
         else None
       | _ -> None
-    if isAnalyzer mi then getAnalyzerFromMemberInfo mi else None
+
+    match isAnalyzer mi with
+    | Some analyzerAttribute ->
+        match getAnalyzerFromMemberInfo mi with
+        | Some analyzer -> Some (analyzerAttribute.Name, analyzer)
+        | None -> None
+    | None -> None
 
   let internal analyzersFromType (t: Type) =
     let asMembers x = Seq.map (fun m -> m :> MemberInfo) x
@@ -61,7 +66,7 @@ module Client =
     |> Seq.choose analyzerFromMember
     |> Seq.toList
 
-  let registeredAnalyzers: ConcurrentDictionary<string, Analyzer list> = ConcurrentDictionary()
+  let registeredAnalyzers: ConcurrentDictionary<string, (string * Analyzer) list> = ConcurrentDictionary()
 
   ///Loads into private state any analyzers defined in any assembly
   ///matching `*Analyzer*.dll` in given directory (and any subdirectories)
@@ -97,9 +102,21 @@ module Client =
       0,0
 
   ///Runs all registered analyzers for given context (file).
-  ///Returns list of messages
+  ///Returns list of messages. Ignores errors from the analyzers
   let runAnalyzers (ctx: Context) : Message list =
     let analyzers = registeredAnalyzers.Values |> Seq.collect id
     analyzers
-    |> Seq.collect (fun analyzer -> analyzer ctx)
+    |> Seq.collect (fun (analyzerName, analyzer) -> try analyzer ctx with error -> [ ])
+    |> Seq.toList
+
+  /// Runs all registered analyzers for given context (file).
+  /// Returns list of results per analyzer which can ei
+  let runAnalyzersSafely (ctx: Context) : AnalysisResult list =
+    let analyzers = registeredAnalyzers.Values |> Seq.collect id
+    analyzers
+    |> Seq.map (fun (analyzerName, analyzer) ->
+        {
+          AnalyzerName = analyzerName
+          Output = try Ok (analyzer ctx) with error -> Result.Error error
+        })
     |> Seq.toList